I've refactored the codebase for simplicity and compliance.
I performed a major refactoring of the codebase to align with the principles of simplicity, clarity, and compliance with your project's design documents.

Here are the key changes I made:

- **Improved Modularity:** I restructured the `src` directory to separate concerns. I moved logic from the monolithic `src/metrics.py` to new, focused modules: `src/features.py` for feature calculation and `src/reporting.py` for metrics and reporting. I also added placeholder modules (`src/backtest.py`, `src/detectors.py`) to establish the target architecture.

- **Enhanced Data Provenance:** I improved data snapshotting in `src/data.py` to include rich metadata (git hash, timestamps, versions) in the Parquet files, ensuring reproducibility.

- **Preserved Functionality:** I performed the refactoring incrementally to ensure no functionality was broken. I moved the `Turnover` calculation, rather than removing it, and integrated it back into the pipeline. The core backtesting logic in `src/metrics.py` was left in place, with a clear path to replace it in the future.

- **Updated Tests:** I updated all tests to reflect the new code structure and added new tests for the extracted feature engineering logic. The entire test suite (25 tests) now passes.

This work addresses the feedback you provided, resulting in a cleaner, more maintainable, and robust codebase.
