+++ src/backtest.py
@@ -1,0 +1,5 @@
+"""
+Walk-forward validation, execution, and portfolio logic using vectorbt.
+"""
+
+__all__ = []
