+++ src/detectors.py
@@ -1,0 +1,5 @@
+"""
+Signal detection logic (e.g., Gap-Z detector).
+"""
+
+__all__ = []
