+++ tests/test_metrics.py
@@ -9,8 +9,8 @@ import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal, assert_series_equal
 
-from src.metrics import (
-    aggregate_portfolio_returns,
+from src.metrics import aggregate_portfolio_returns
+from src.reporting import (
     calculate_benchmark_metrics,
     calculate_oos_is_ratio,
     calculate_per_stock_metrics,
